Move to pragma once. Move update code to InputManager.
