+++ pcsx2/PAD/Linux/InputManager.cpp
@@ -15,7 +15,7 @@
 
 #include "InputManager.h"
 #include "Device.h"
-
+#include "keyboard.h"
 #include "state_management.h"
 
 #ifdef SDL_BUILD
@@ -33,13 +33,52 @@ InputDeviceManager::~InputDeviceManager()
 	device_manager->devices.clear();
 }
 
-void InputDeviceManager::Update()
+// Needs to be moved to individual device code, as does the keyboard input.
+void PollForJoystickInput(int cpad)
 {
+	int index = Device::uid_to_index(cpad);
+	if (index < 0) return;
+
+	auto& gamePad = device_manager->devices[index];
+
+	gamePad->UpdateDeviceState();
+
+	for (u32 i = 0; i < MAX_KEYS; i++)
+	{
+		s32 value = gamePad->GetInput((gamePadValues)i);
+		if (value != 0)
+			g_key_status.press(cpad, i, value);
+		else
+			g_key_status.release(cpad, i);
+	}
 }
 
-/**
- * Find every interesting devices and create right structure for them(depend on backend)
- **/
+void InputDeviceManager::Update()
+{
+	// Poll keyboard/mouse event. There is currently no way to separate pad0 from pad1 event.
+	// So we will populate both pad in the same time
+	for (u32 cpad = 0; cpad < GAMEPAD_NUMBER; cpad++)
+	{
+		g_key_status.keyboard_state_acces(cpad);
+	}
+	UpdateKeyboardInput();
+
+	// Get joystick state + Commit
+	for (u32 cpad = 0; cpad < GAMEPAD_NUMBER; cpad++)
+	{
+		g_key_status.joystick_state_acces(cpad);
+
+		PollForJoystickInput(cpad);
+
+		g_key_status.commit_status(cpad);
+	}
+
+	Pad::rumble_all();
+}
+
+/*
+ * Find and set up joysticks, potentially other devices.
+ */
 void EnumerateDevices()
 {
 #ifdef SDL_BUILD
