+++ pcsx2/PAD/Linux/linux.cpp
@@ -57,26 +57,6 @@ void _PADclose()
 	device_manager->devices.clear();
 }
 
-void PollForJoystickInput(int cpad)
-{
-	int index = Device::uid_to_index(cpad);
-	if (index < 0)
-		return;
-
-	auto& gamePad = device_manager->devices[index];
-
-	gamePad->UpdateDeviceState();
-
-	for (u32 i = 0; i < MAX_KEYS; i++)
-	{
-		s32 value = gamePad->GetInput((gamePadValues)i);
-		if (value != 0)
-			g_key_status.press(cpad, i, value);
-		else
-			g_key_status.release(cpad, i);
-	}
-}
-
 void PADupdate(int pad)
 {
 #ifndef __APPLE__
@@ -94,26 +74,7 @@ void PADupdate(int pad)
 
 	// Actually PADupdate is always call with pad == 0. So you need to update both
 	// pads -- Gregory
-
-	// Poll keyboard/mouse event. There is currently no way to separate pad0 from pad1 event.
-	// So we will populate both pad in the same time
-	for (u32 cpad = 0; cpad < GAMEPAD_NUMBER; cpad++)
-	{
-		g_key_status.keyboard_state_acces(cpad);
-	}
-	UpdateKeyboardInput();
-
-	// Get joystick state + Commit
-	for (u32 cpad = 0; cpad < GAMEPAD_NUMBER; cpad++)
-	{
-		g_key_status.joystick_state_acces(cpad);
-
-		PollForJoystickInput(cpad);
-
-		g_key_status.commit_status(cpad);
-	}
-
-	Pad::rumble_all();
+	device_manager->Update();
 }
 
 void PADconfigure()
