+++ pcsx2/PAD/Linux/wx_dialog/JoystickConfiguration.h
@@ -15,9 +15,6 @@
 
 #pragma once
 
-#ifndef __JOYSTICKCONFIGURATION_H__
-#define __JOYSTICKCONFIGURATION_H__
-
 #include <wx/wx.h>
 
 #include "../Global.h"
@@ -43,5 +40,3 @@ public:
 	JoystickConfiguration(int, bool, wxWindow*);
 	void InitJoystickConfiguration();
 };
-
-#endif // __JOYSTICKCONFIGURATION_H__
